+++ module/combat.js
@@ -498,6 +498,13 @@ export class HeroSystem6eCombat extends Combat {
             await ChatMessage.create(chatData)
         }
 
+        // Some attacks include a DCV penalty which was added as an ActiveEffect.
+        // At the beginning of our turn we make sure that AE is deleted.
+        const removeOnNextPhase = combatant.actor.effects.filter(o=> o.flags.nextPhase && o.duration.startTime < game.time.worldTime);
+        for (const ae of removeOnNextPhase) {
+            await ae.delete();
+        }
+
     }
 
     /**